+++ src/test/java/kitchenpos/application/InMemoryOrderRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.application;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
