+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -5,8 +5,8 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.order.domain.OrderType;
 import kitchenpos.order.application.OrderService;
 import org.junit.jupiter.api.BeforeEach;
